convert `bittensor.core.registration.py` module to `bittensor.core.registration` sub-package
